test-requirements.txt
python-openstackclient>=6.3.0 # Apache-2.0
openstacksdk>=0.10.0 # Apache-2.0
stestr>=2.0.0 # Apache-2.0
tempest>=10 # Apache-2.0
testtools>=1.4.0 # MIT
